Enemy Health Bars

+++ First Person Shooter/Assets/Scripts/DestroyInOneHit.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DestroyInOneHit : MonoBehaviour, IDamageable
 {
-    public void DealDamage(int damage)
+    public void DealDamage(float damage)
     {
         Destroy(gameObject);
     }
