+++ First Person Shooter/Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageable
 {
-    void DealDamage(int damage);
+    void DealDamage(float damage);
 }
